Disable auth implementation temporarily
The entire authentication area, which includes providers and tokens creation, has been temporarily commented out in both auth.ts and route.ts files. This is to allow for testing the functionality of other areas of the app without triggering authentication to speed up development. The disabled auth implementation will be reintroduced once the testing of other functionality is complete.